ml/ris/gastro/requirements.txt
certifi==2018.1.18
dill==0.2.7.1
numpy==1.14.2
pandas==0.22.0
jupyter==1.0.0
configparser==3.5.0
texttable==1.2.1
nltk==3.2.5
scikit-learn==0.19.1
matplotlib==2.1.2
tpot==0.9.2
keras==2.1.4
tensorflow==1.5.0
gensim==3.4.0
wordcloud
pillow

